requirements.txt
python-dotenv==0.17.1
SQLAlchemy==1.4.15
psycopg2-binary
numpy
matplotlib